rename: sharedDepUtils.js -> runtimeBinaryCheck.js
