+++ apps/vs-code-designer/src/test/ui/runtimeBinaryCheck.js
